pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "intugle"
version = "1.1.0"
authors = [
    { name="Intugle", email="hello@intugle.ai" },
]
description = "A GenAI-powered Python library for building semantic layers."
readme = "README.md"
keywords = ["data", "semantic layer", "genai", "llm", "data profiling", "link prediction", "sql generation"]
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules"
]

dependencies = [
    "langchain-community>=0.3.21,<1.0.0",
    "langchain-openai>=0.3.28,<1.0.0",
    "langgraph>=0.6.4",
    "nltk>=3.9.1",
    "numpy<=2.3.0",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.116.1",
    "matplotlib>=3.10.5",
    "mcp[cli]>=1.12.4",
    "networkx>=3.4.2",
    "pandas>=2.2.2",
    "pyaml>=25.7.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pyfunctional>=1.5.0",
    "python-dotenv>=1.1.1",
    "symspellpy>=6.9.0",
    "trieregex>=1.0.0",
    "xgboost>=3.0.4",
    "pyyaml>=6.0.2",
    "duckdb>=1.3.2",
    "scikit-learn==1.7.1",
    "langchain[anthropic,google-genai,openai]>=0.3.27,<1.0.0",
    "qdrant-client>=1.15.1",
    "rich>=14.1.0",
    "aiohttp>=3.9.5",
    "aiofiles>=23.2.1",
    "tavily-python>=0.1.11",
]

[project.optional-dependencies]
snowflake = [
    "snowflake-snowpark-python[pandas]>=1.12.0",
    "sqlglot>=27.20.0",
]
databricks = [
    "databricks-sql-connector>=4.1.3",
    "pyspark>=3.5.0,<4.0.0",
    "sqlglot>=27.20.0",
]
postgres = [
    "asyncpg>=0.30.0",
    "sqlglot>=27.20.0",
]
sqlserver = [
    "mssql-python>=0.13.1",
    "sqlglot>=27.20.0",
]

streamlit = [
    "streamlit==1.50.0",
    "pyngrok==7.4.0",
    "python-dotenv==1.1.1",
    "xlsxwriter==3.2.9",
    "plotly",
    "graphviz"
]


[project.urls]
"Homepage" = "https://github.com/Intugle/data-tools"
"Bug Tracker" = "https://github.com/Intugle/data-tools/issues"

[project.scripts]
intugle-mcp = "intugle.mcp.server:main"
intugle-streamlit = "intugle.cli:run_streamlit_app"

[dependency-groups]
test = [
    "langchain-core>=0.3.74",
    "langchain-mcp-adapters>=0.1.9",
    "pytest",
    "pytest-asyncio>=1.1.0",
    "pytest-mock",
    "sqlglot>=27.20.0",
]
lint = ["ruff"]
dev = [
    "asyncpg>=0.30.0",
    "databricks-sql-connector>=4.1.3",
    "ipykernel>=6.30.1",
    "langfuse==2.60.4",
    "mssql-python>=0.13.1",
    "pysonar>=1.2.0.2419",
    "pyspark>=3.5.0,<4.0.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.15.1",
    "ruff>=0.13.2",
    "snowflake-snowpark-python[pandas]>=1.39.1",
    "sqlglot>=27.20.0",
    "twine>=6.1.0",
]

[tool.ruff]
src = ["src"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"intugle" = ["**/*.yaml", "**/*.txt", "**/*.pkl", "mcp/semantic_layer/prompts/*.md", "streamlit_app/**/*"]

[tool.pytest.ini_options]
markers = [
    "mcp: marks tests that require the MCP server to be running",
]

[tool.sonar]
projectKey = "Intugle-Data-Tools"
projectName = "Intugle Data Tools"
sources = "src"
python.version = "3.10"
python.coverage.reportPaths = "coverage.xml"
